Implement a pseudo refresh function Haskell-side.
This is practical only for the message shown when the file is new. If
the file exists already, its content is hidden until the next keypress.
Then it uses the original refresh function.

To query how many lines there are to clear, another FFI call is added.
Probably it should expose the whole struct editorConfig E.